+++ src/main/java/org/codehaus/groovy/runtime/StringGroovyMethods.java
@@ -4539,5 +4539,3 @@ public class StringGroovyMethods extends DefaultGroovyMethodsSupport {
         return BalancedGroup.find(self, openRegex, closeRegex, options);
     }
 }
-
-
